Make the use of the CastAsCall fixture configurable
In a subsequent patch the CastAsCall fixture will need to *not* be
used in new functional tests in order to avoid NoValidHost
exceptions from the scheduler select_destination() cast/call ending up
in the API as 500 errors. This patch adds a USE_CAST_AS_CALL flag,
defaulted to True, that allows test classes to disable the CastAsCall
fixture if they don't want it.

Change-Id: I13deb7cb02414ffbf8758911a5b69baa69cd1078